requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
idna==3.10
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.1
pydantic-core==2.27.1
pyparsing==3.2.0
pytube==15.0.0
requests==2.32.3
rsa==4.9
tqdm==4.67.0
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yt-dlp==2024.11.18
